Update g.analyse.perday.R
In the last qqq window in the g.analyse.perday function, qqq1 is higher than qqq2, which makes no sense. In this case, the function does not run an error or warning message (even with do.parallel = F), but the file is just not saved into the ms2.out directory.

I set a conditional such as: if (qqq1 > qqq2) next().

It works ok with my files.