Function to read temperature from movisens files
Movisens store a file per sensor, so that the temperature is stored in a different file than accelerations. Additionally, they work at different sample rates, so the resample function has been used to make them compatible. This function reads the temperature and adds it to the matrix where accelerations are stored.