fix sleeponset wakeup writing error on first day
Only use the startTimeRecord method if the start of the block being passed into HDCZA started after noon (ie, the whole block was less than 24h). Otherwise, this will result in an incorrect time if (j == 1) and the recording started before noon (which was the error that I was seeing).