fix error on writing sleeponset and wakeup times on first day
The code assumed that on the first day of recording, the start time of the file was equivalent to the start of the block of data sent into the HDCZA algorithm. This was only true when the recording started after noon. I've updated the code to only use this method when the recording starts after noon. If the recording starts before then, then it can use the normal offset from noon. To do this check I used the start index on the angle data that is sent into the HDCZA algorithm. IF, the start index is 1, then there is less than a full 24h data chunk being sent into HDCZA and the startTimeRecord method should be used.