update non-wear explanation
The non-wear explanation did not match the code functionality. 

The vignette stated that non-wear was identified when:
'A block was classified as non-wear time if the standard deviation of the 60 minute window was less than 13.0 mg ($1 mg = 0.00981 m·s^−2$) for at least two out of the three axes or if the value range, for at least two out of three axes, was less than 50 mg.'

However, when looking at the code block (line 634 / 635 in g.getmeta.R), we see that the standard deviation criteria and and range criteria must be met for the axis in question to be labelled as non-wear. Then, later in the code, if there are at least two axes labelled as non-wear, the algorithm will classify the ws2 (15min) block was non-wear.

I have updated the documentation in the vignette to reflect this.