c() is NULL, which cannot be put into a vector; NA can be.
c() is equivalent to NULL, which is causing problems in the following line in `g.analyse.perfile.R`:

filesummary[vi:(vi+3)] = c(InterdailyStability, IntradailyVariability,
                               IVIS_windowsize_minutes, IVIS_epochsize_seconds)

The thing is, NULL is not allowed in a vector, and so it's getting ignored, and we end up with a vector size of 3 instead of 4. This causes a warning when Part2 is run:

"number of items to replace is not a multiple of replacement length"

NA doesn't have this problem and can be used to represent a missing value.