fix test file generation to allow for sessions of different length
create_test_acc_csv() was failing whenever Nmin was such that one of the sleep blocks started just before the end of the recording (and so the end of the sleep block would have been after the recording end). Same issue for non-wear periods and activity periods.

This fix ensures that no sleep block (or non-wear block, or activity block) ends after the end of the recording. If a block is too long to fully fit within a recording, it gets truncated.