Updates to handle corrupt gt3x files #873 and simplification of the code
Updates to handle corrupt gt3x files #873 the code no longer errors when it encounters a corrupt file but goes to next file and logs the encounter of a corrupt file in g.part2.report. Further, and g.getmeta now also uses inspectfileobject, which allows for further simplification of the code.