specify environment for exists()
R uses lexical scoping, not dynamic scoping. So when it creates a new environment to call the function exists(), by default this environment is a child of the base environment. The base env doesn't have any of the variables in expectedArgs defined, so lapply(expectedArgs, FUN = exists) will always return a list of FALSEs, unless we explicitly specify the environment we want for function exists().

Because of this,  extract_params() was called all the time, unnecessarily.

Also, list(...) will not contain any explicitly named parameters, so no need to check for that.