POSIXct stores time in seconds since 1970-01-01, in UTC
Parameter tz doesn't influence this internally stored value, only the textual representation when the timestamp is printed out.

So if we start off with a unix timestamp, it doesn't make sense to first convert it to POSIXct(, tz = configtz) and then convert the result back to unix seconds and then POSIXct(, tz = desiredtz). Use desiredtz right away, since tz is just a label that gets added to the same unix timestamp.