make skip of report generation less verbose
we now have report turned on by default for all parts which will generated a lot of console messages if the user only wants to run a few parts, I think this clutters the console unnecessarily. So, i am now turning these off when the report is not generated. If the report is generated all works as normal.