don't convert strings to factors
stringsAsFactors=TRUE was set in response to R changing the default for this parameter from TRUE to FALSE.

But we don't expect any non-numeric values in these data files, and if there's something wrong with the file and it contains a character string, it's better to end up with a character column than a factor column. Factors can be mistaken for numbers, and this conversion could end up undetected, even though we'd end up with very unexpected numeric values.