fix number of rows skipped as header
1. Always skip (rmc.firstrow.acc - 1) rows, not rmc.firstrow.acc like we did when length(rmc.firstrow.header) == 0, because otherwise the first row of data is lost.

2. Set header parameter of data.table::fread call to "auto" (this way, IFF every non-empty field on the first data line is of type character, this line will be read in as column names).  Don't just assume that there is a header with column names, people might have set rmc.firstrow.acc to point to the row after this header. And especially if rmc.firstrow.header is set, then rmc.firstrow.acc is better be set to point to the actual first row of acc data, since data.table::fread won't know how to deal with an arbitrary length/structure header block.

For the case of rmc.firstrow.acc == 2, we used to mistakenly skip 2 rows of data because we were setting skip = rmc.firstrow.acc (== 2), and were also calling data.table::fread with header = TRUE (so using a row of data as if it were column names), even though the column names were most likely on row 1 which we had skipped anyway.

And for the case when rmc.firstrow.header was set, we were most likely losing one row of data, because then rmc.firstrow.acc was most likely pointing to the first row of the actual acc data, not to any header info.

And if rmc.firstrow.acc does point to a row of column names, then header = "auto" can handle that just fine.