startpage++ not always needed for AD_HOC_CSV
only do startpage++ for FORMAT$AD_HOC_CSV if rmc.firstrow.acc is pointing at a row containing column names.

Otherwise if there is no header with column names, or if rmc.firstrow.acc was pointing at the row after this header, we were skipping the first row of data for no reason.