don't lose data from the last block even if short
Changing the logic to be similar to that for other monitor types.

We only want to get rid of a short chunk of data if this was the first block (because then there's not enough data for meaningful analysis anyway). If it's not the first block, this short amount of data should be used and added at the end of the previosly read blocks.