correct timezones of gt3x timestamps
Note that simply changing the timestamp of a POSIXct object actually changes the timezone and the hh:mm:ss representation, while keeping the same underlying unix timestamp. That's not what we want. We want to change the unix timestamp in such a way that when printed with the configtz timestamp, the time comes out as the same hh:mm:ss value as before.

For instance, we want
"2020-08-26 10:09:00 GMT"
to become
"2020-08-26 10:09:00 EST", not "2020-08-26 06:09:00 EDT", which is what we'd get with as.POSIXct(..., tz=configtz)

An alternative solution would have been to use lubridate::force_tz(), but I wanted to avoid pulling in lubridate.

Note that another solution, suggested here:
https://stackoverflow.com/questions/15575713/modifying-timezone-of-a-posixct-object-without-changing-the-display

as.numeric(as.POSIXct(as.numeric(timestamp), origin = as.POSIXct("1970-01-01", tz=configtz), tz=configtz))

actually doesn't work with daytime savings zones. Since "1970-01-01" is in the winter timezone, any timestamp that is in the summer timezone will end up 1 hour off (but winter timestamps will be fine).