Allocate metalong light vars even if no temperature
Allocate metalong light variables if needed, even if temperature is not available. For currently supported monitors this will only come up if temperature is not used because it's unrealistically high.

Also remove warning suppression. It came from an older version of code and it's not clear that the warning it was meant to suppress is still relevant.