rephrase documentation for use_trycatch_serial #1308
I have rephrase this as it needs to be clear that do.parallel already uses trycatch, so this is not some kind of special new feature but only relevant in a very specific use-case.
Note that "GGIR will stop execution immediately when an error  occurs as usual." is incorrect. The usual scenario is that GGIR used in parallel where error are caught by trycatch.